+++ source/isaaclab/test/sim/test_spawn_meshes.py
@@ -117,110 +117,25 @@ def test_spawn_sphere(sim):
     assert prim.GetPrimTypeInfo().GetTypeName() == "Mesh"
 
 
+@pytest.mark.parametrize("resolution", [(1, 1), (3, 2)])
+def test_spawn_square(sim, resolution):
+    """Test spawning of UsdGeomMesh as a square prim."""
+    # Spawn square
+    cfg = sim_utils.MeshSquareCfg(size=1.0, resolution=resolution)
+    prim = cfg.func("/World/Square", cfg)
+
+    # Check validity
+    assert prim.IsValid()
+    assert sim.stage.GetPrimAtPath("/World/Square").IsValid()
+    assert prim.GetPrimTypeInfo().GetTypeName() == "Xform"
+    # Check properties
+    prim = sim.stage.GetPrimAtPath("/World/Square/geometry/mesh")
+    assert prim.GetPrimTypeInfo().GetTypeName() == "Mesh"
+
+
 """
 Physics properties.
 """
-
-
-def test_spawn_cone_with_deformable_props(sim):
-    """Test spawning of UsdGeomMesh prim for a cone with deformable body API."""
-    # Spawn cone
-    cfg = sim_utils.MeshConeCfg(
-        radius=1.0,
-        height=2.0,
-        deformable_props=sim_utils.DeformableBodyPropertiesCfg(deformable_enabled=True),
-    )
-    prim = cfg.func("/World/Cone", cfg)
-
-    # Check validity
-    assert prim.IsValid()
-    assert sim.stage.GetPrimAtPath("/World/Cone").IsValid()
-
-    # Check properties
-    # Unlike rigid body, deformable body properties are on the mesh prim
-    prim = sim.stage.GetPrimAtPath("/World/Cone/geometry/mesh")
-    assert prim.GetAttribute("physxDeformable:deformableEnabled").Get() == cfg.deformable_props.deformable_enabled
-
-
-def test_spawn_cone_with_deformable_and_mass_props(sim):
-    """Test spawning of UsdGeomMesh prim for a cone with deformable body and mass API."""
-    # Spawn cone
-    cfg = sim_utils.MeshConeCfg(
-        radius=1.0,
-        height=2.0,
-        deformable_props=sim_utils.DeformableBodyPropertiesCfg(deformable_enabled=True),
-        mass_props=sim_utils.MassPropertiesCfg(mass=1.0),
-    )
-    prim = cfg.func("/World/Cone", cfg)
-
-    # Check validity
-    assert prim.IsValid()
-    assert sim.stage.GetPrimAtPath("/World/Cone").IsValid()
-    # Check properties
-    prim = sim.stage.GetPrimAtPath("/World/Cone/geometry/mesh")
-    assert prim.GetAttribute("physics:mass").Get() == cfg.mass_props.mass
-
-    # check sim playing
-    sim.play()
-    for _ in range(10):
-        sim.step()
-
-
-def test_spawn_cone_with_deformable_and_density_props(sim):
-    """Test spawning of UsdGeomMesh prim for a cone with deformable body and mass API.
-
-    Note:
-        In this case, we specify the density instead of the mass. In that case, physics need to know
-        the collision shape to compute the mass. Thus, we have to set the collider properties. In
-        order to not have a collision shape, we disable the collision.
-    """
-    # Spawn cone
-    cfg = sim_utils.MeshConeCfg(
-        radius=1.0,
-        height=2.0,
-        deformable_props=sim_utils.DeformableBodyPropertiesCfg(deformable_enabled=True),
-        mass_props=sim_utils.MassPropertiesCfg(density=10.0),
-    )
-    prim = cfg.func("/World/Cone", cfg)
-
-    # Check validity
-    assert prim.IsValid()
-    assert sim.stage.GetPrimAtPath("/World/Cone").IsValid()
-    # Check properties
-    prim = sim.stage.GetPrimAtPath("/World/Cone/geometry/mesh")
-    assert prim.GetAttribute("physics:density").Get() == cfg.mass_props.density
-    # check sim playing
-    sim.play()
-    for _ in range(10):
-        sim.step()
-
-
-def test_spawn_cone_with_all_deformable_props(sim):
-    """Test spawning of UsdGeomMesh prim for a cone with all deformable properties."""
-    # Spawn cone
-    cfg = sim_utils.MeshConeCfg(
-        radius=1.0,
-        height=2.0,
-        mass_props=sim_utils.MassPropertiesCfg(mass=5.0),
-        deformable_props=sim_utils.DeformableBodyPropertiesCfg(),
-        visual_material=sim_utils.PreviewSurfaceCfg(diffuse_color=(0.0, 0.75, 0.5)),
-        physics_material=sim_utils.DeformableBodyMaterialCfg(),
-    )
-    prim = cfg.func("/World/Cone", cfg)
-
-    # Check validity
-    assert prim.IsValid()
-    assert sim.stage.GetPrimAtPath("/World/Cone").IsValid()
-    assert sim.stage.GetPrimAtPath("/World/Cone/geometry/material").IsValid()
-    # Check properties
-    # -- deformable body
-    prim = sim.stage.GetPrimAtPath("/World/Cone/geometry/mesh")
-    assert prim.GetAttribute("physxDeformable:deformableEnabled").Get() is True
-
-    # check sim playing
-    sim.play()
-    for _ in range(10):
-        sim.step()
 
 
 def test_spawn_cone_with_all_rigid_props(sim):